--- src/mastra/tools/workflow/index.ts
@@ -1,7 +1,0 @@
-// Export all workflow tool adapters
-export { default as taskToolAdapter } from './taskToolAdapter';
-export { default as taskHistoryToolAdapter } from './taskHistoryToolAdapter';
-export { default as versionToolAdapter } from './versionToolAdapter';
-
-// Export the factory function for creating new adapters
-export { createToolAdapter } from './createToolAdapter';
